Create an author key
Instead of using Hugo's "urlize" feature to transform names into keys,
create the key in Perl code and store it in the author object
and article instead.

This avoids issues where urlize transforms an author name differently to
a Perl regex, such as with multi-byte unicode letters in author names.