Quoting every program name doesn't format too well
The styling makes the text a bit hard to read, as the command names are
so frequent. This removes most of the backtick quoting so that generaly
only variables and inline commands are shown as code.

This is a subjective thing though, we can drop this commit if needs be.