Pull in latest articles with a jQuery call (#264)
* Ignore generating the site in _site, and run logs

* Don't paginate the pages that declare "exclude_paginate"

So far that's about.md and latest_articles.md, but it
might be more. We can't list all the page names there.

Ideally we'd make a paginator from the pages that don't
have that set. I think the way it is now, some pages
might end up with one (or more) less than it should have.

* Make the latest articles bit some jQuery loading
of a generated page.

* Create the latest articles snippet page

* ignore built/

* Improve the Makefile by merging what I've had elsewhere