Fix for #243
Media widths below 768px result in no padding on the `#content section.entry-content div.container` element.
The footer content has a left margin defined as 20px, which IMO is too much for the article content itself.

An absolute value between 5px and 15px may be appropriate, or 0.5 to 2%?
I tried to find other elements to base the value on. The 9.5px is also the padding applied to `<pre>` elements.

NB I believe #243 is not iOS specific - I witness the same in Chrome on Android.