Fix comment about '0' version number in Map::Tube HOWTO
It turns out that `'0'` *doesn't* specify the latest version: it
specifies that *any* version is ok.  That the latest version is usually
installed is only a side-effect of the given module not having been
installed beforehand.  This change fixes the potentially confusing and
misleading comment about `'0'` giving the latest version.