Update bin/deploy to use latest hugo
I removed a lot of the logic which was meant to prevent version
mismatches in the case where releases could be done either a locally
installed hugo or a Docker image. Since I'm the only one who had
deployed this site in years, I'm switching it to use purely Docker and
pinning it to today's latest available image.

Ideally this would deploy from CI anyway and how hugo gets installed
wouldn't be an issue.