Remove ZeroClipboard library
I happened to stumble across this file and wondered what it was and if
it's still useful.  As the description says

> The ZeroClipboard library provides an easy way to copy text to the
> clipboard using an invisible Adobe Flash movie and a JavaScript
> interface.

Since this is Adobe Flash it's likely *very* out of date and probably a
security issue.  Fortunately, the file wasn't used anywhere else in the
code (i.e. `git grep -i zeroclipboard` didn't give any results).  The
file was added in the [initial project commit](921dc6e2) (found via `git
log -S` and it's hard to see if it's been used anywhere since then.

Because this file doesn't seem to be used anywhere, and because the
library uses a deprecated technology, I think it's a good idea to remove
the file from the repository, hence this commit.