Add canonicalUrl key to new article metadata
I noticed a while ago that Dave Cross had used the `canonicalURL` key in
one of his articles and have hence also used this key in my recent
articles to point to where the article was first published.  To avoid
having to think of adding this key each time myself, I thought it might
be an idea to have this key appear as part of the default new article
metadata, hence the change implemented in this commit.

Note that the key I've added uses a different capitalisation scheme to
that which I mentioned above.  Grepping through the source shows that
`layouts/partials/header.html` uses the value of `.Params.canonicalUrl`
to decide whether or not to include this information in the output HTML
header metadata.  Therefore, I thought it more consistent to use
`canonicalUrl` as opposed to `canonicalURL` for the default metadata
key.

I've also found that using an empty value for `canonicalUrl` provides
the default behaviour that currently exists when this key is not
mentioned in the article metadata.  In other words, setting
`canonicalUrl` to the empty string by default doesn't change the current
behaviour of how HTML is generated on the site.