Space out new article metadata after canonicalUrl addition
The `canonicalUrl` key is rather long and hence there wasn't any space
between the new key and the colon separating it from its value.  This
change is a purely formatting change to space out the initial article
metadata so that it's not as visually cramped and hence easier to read
and understand.